Revert "wip: remove type cast int8 on arg_types"
This reverts commit 6878a93bfbc8c4a87f5dd36d5a1ed13381760d85.

+++ packages/pg-meta/src/sql/functions.ts
@@ -89,7 +89,7 @@ from
           oid,
           unnest(arg_modes) as mode,
           unnest(arg_names) as name,
-          unnest(arg_types) as type_id,
+          unnest(arg_types)::int8 as type_id,
           unnest(arg_has_defaults) as has_default
         from
           functions
